Remove unused task
